+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.35-SNAPSHOT
 
+### Breaking changes
+- [UI] Now `SearchPlace` constructor requires additional property `id`. `copy()` function's signature has also been changed.
+
 ### Mapbox dependencies
 - Search Native SDK `0.57.0`
 - Common SDK `22.1.0-beta.1`